refactor: introduce cdk folder
