+++ projects/movies/src/app/pages/movie-detail-page/movie-detail-page.adapter.ts
@@ -10,11 +10,11 @@ import {
 } from '../../data-access/api/resources/movie.resource';
 import { transformToMovieDetail } from './selection/client-movie-detail.mapper';
 import { getActions } from '../../shared/rxa-custom/actions';
-import { infiniteScrolled } from '../../shared/utils/infinite-scroll/infinite-scrolled';
+import { infiniteScrolled } from '../../shared/cdk/infinite-scroll/infinite-scrolled';
 import { MovieDetail } from './selection/movie-detail.model';
 import { TMDBMovieModel } from '../../data-access/api/model/movie.model';
 import { TMDBMovieCastModel } from '../../data-access/api/model/movie-credits.model';
-import { InfiniteScrollState } from '../../shared/utils/infinite-scroll/paginate-state.interface';
+import { InfiniteScrollState } from '../../shared/cdk/infinite-scroll/paginate-state.interface';
 
 export interface MovieDetailPageModel {
   loading: boolean;
