+++ projects/movies/src/app/shared/cdk/element-visibility/observe-element-visibility.ts
